Fixed Probability immediately set to 0 issue - Fixed bug where constInvTemp() was return an int. Since inv temp is usually a double and less then 1, function was returning 0, causing probability of learning to be 0 all the time - Todo: program behavior still does not match previous implementation. More debugging to come

+++ MultiNightBar.hpp
@@ -109,7 +109,7 @@ public:
     void updatePrevG(double newG);
 
     // constant temperature function - returns the inverse of temperature value to save computation
-    int constInvTemp(int epochNumber);
+    double constInvTemp(int epochNumber);
 
     // creates ofstreams for each of the loggers based on the string path provided
     void setupLoggers();
